Add files via upload
Completed graph features and added "Circuit Connection" function at CTC tab. 
Now several netlist and touchstone files can be loaded. 
To simulate a single netlist it must first be loaded then selected from the "Netlist Files" (it is selected by default one it is loaded), them set all simulation parameters and hit run.
To simulate a single touchstone or a netlist-touchstone-netlist the files to use for the simulation must be selected at the CTC tab listboxes, and then hit simulate. Note that the CTC feature is limited for only 2 ports files.